kite-python/kite_emr/requirements.txt
boto3==1.3.1
pyyaml==3.11
spooky==2.0.0
chardet==2.3.0
